[dev-launcher][android] move authentication fns to dev-launcher
